+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- `useEffect` now call its unmount function when its dependencies change, just like React hooks.
+
 ## [0.2.0]
 ### Added
 - Added validateProps component API option for hooked components. `validateProps(props) -> (false, message: string) | true`